@@ -4,15 +4,15 @@ description: A common library for charting packages.
 homepage: https://github.com/google/charts
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.19.5 <4.0.0'
 
 dependencies:
-  collection: ^1.14.5
-  intl: ">=0.15.2 < 0.18.0"
-  logging: ^1.0.2
-  meta: ^1.1.1
-  vector_math: ^2.0.8
+  collection: any
+  intl: any
+  logging: any
+  meta: any
+  vector_math: any
 
 dev_dependencies:
-  mockito: ^5.0.0
-  test: ^1.5.3
+  mockito: any
+  test: any
