@@ -1,10 +1,10 @@
 name: charts_flutter
-version: 0.12.0
+version: 1.0.0
 description: Material Design charting library for flutter.
 homepage: https://github.com/google/charts
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.19.5 <4.0.0'
   flutter: '>=2.5.0'
 
 dependencies:
@@ -13,20 +13,19 @@ dependencies:
   #
   # The pub version of charts_flutter will point to the pub version of charts_common.
   # The latest pub version is commented out and shown below as an example.
-  charts_common: ^0.12.0
-  # charts_common:
-  #   path: ../charts_common/
-  collection: ^1.14.5
+  charts_common:
+    path: ../charts_common/
+  collection: any
   flutter:
     sdk: flutter
-  intl: ">=0.15.2 < 0.18.0"
-  logging: ^1.0.2
-  meta: ^1.1.1
+  intl: any
+  logging: any
+  meta: any
 
 
 dev_dependencies:
-  mockito: ^5.0.0
-  build_runner: ^1.11.0
+  mockito: any
+  build_runner: any
   flutter_test:
     sdk: flutter
-  test: ^1.3.0
+  test: any
