@@ -156,8 +156,7 @@ class LinearScaleFunction {
         case StepSizeType.autoDetect:
           final minimumDetectedDomainStep =
               domainInfo.minimumDetectedDomainStep.toDouble();
-          if (minimumDetectedDomainStep != null &&
-              minimumDetectedDomainStep.isFinite) {
+          if (minimumDetectedDomainStep.isFinite) {
             scalingFactor = viewportSettings.scalingFactor *
                 (rangeDiff /
                     (domainDiff +
