@@ -519,14 +519,12 @@ class LineRenderer<D> extends BaseCartesianRenderer<D> {
     // Animate out lines that don't exist anymore.
     _seriesLineMap.forEach((String key, List<_AnimatedElements<D>> elements) {
       for (var element in elements) {
-        if (element.lines != null) {
-          for (var line in element.lines) {
-            if (!_currentKeys.contains(line.key)) {
-              line.animateOut();
-            }
+        for (var line in element.lines) {
+          if (!_currentKeys.contains(line.key)) {
+            line.animateOut();
           }
         }
-        if (element.areas != null) {
+              if (element.areas != null) {
           for (var area in element.areas!) {
             if (!_currentKeys.contains(area.key)) {
               area.animateOut();
@@ -957,13 +955,11 @@ class LineRenderer<D> extends BaseCartesianRenderer<D> {
             .map<_AreaRendererElement<D>>((_AnimatedArea<D> animatingArea) =>
                 animatingArea.getCurrentArea(animationPercent))
             .forEach((area) {
-          if (area != null) {
-            canvas.drawPolygon(
-                clipBounds: _getClipBoundsForExtent(area.positionExtent),
-                fill: area.areaColor ?? area.color,
-                points: area.points.toPoints());
-          }
-        });
+          canvas.drawPolygon(
+              clipBounds: _getClipBoundsForExtent(area.positionExtent),
+              fill: area.areaColor ?? area.color,
+              points: area.points.toPoints());
+                });
       }
 
       if (_hasMeasureBounds) {
@@ -975,13 +971,11 @@ class LineRenderer<D> extends BaseCartesianRenderer<D> {
             .map<_AreaRendererElement<D>>((_AnimatedArea<D> animatingBounds) =>
                 animatingBounds.getCurrentArea(animationPercent))
             .forEach((bound) {
-          if (bound != null) {
-            canvas.drawPolygon(
-                clipBounds: _getClipBoundsForExtent(bound.positionExtent),
-                fill: bound.areaColor ?? bound.color,
-                points: bound.points.toPoints());
-          }
-        });
+          canvas.drawPolygon(
+              clipBounds: _getClipBoundsForExtent(bound.positionExtent),
+              fill: bound.areaColor ?? bound.color,
+              points: bound.points.toPoints());
+                });
       }
 
       if (config.includeLine) {
@@ -993,16 +987,14 @@ class LineRenderer<D> extends BaseCartesianRenderer<D> {
             .map<_LineRendererElement<D>>((_AnimatedLine<D> animatingLine) =>
                 animatingLine.getCurrentLine(animationPercent))
             .forEach((line) {
-          if (line != null) {
-            canvas.drawLine(
-                clipBounds: _getClipBoundsForExtent(line.positionExtent!),
-                dashPattern: line.dashPattern,
-                points: line.points!.toPoints(),
-                stroke: line.color,
-                strokeWidthPx: line.strokeWidthPx,
-                roundEndCaps: line.roundEndCaps);
-          }
-        });
+          canvas.drawLine(
+              clipBounds: _getClipBoundsForExtent(line.positionExtent!),
+              dashPattern: line.dashPattern,
+              points: line.points!.toPoints(),
+              stroke: line.color,
+              strokeWidthPx: line.strokeWidthPx,
+              roundEndCaps: line.roundEndCaps);
+                });
       }
     });
 
@@ -1584,12 +1576,10 @@ class _AnimatedElements<D> {
     }
 
     var linesAnimatingOut = true;
-    if (lines != null) {
-      for (final line in lines) {
-        linesAnimatingOut = linesAnimatingOut && line.animatingOut;
-      }
+    for (final line in lines) {
+      linesAnimatingOut = linesAnimatingOut && line.animatingOut;
     }
-
+  
     var boundsAnimatingOut = true;
     if (bounds != null) {
       for (final bound in bounds!) {
@@ -1609,12 +1599,10 @@ class _AnimatedElements<D> {
     }
 
     var linesOverlaySeries = true;
-    if (lines != null) {
-      for (final line in lines) {
-        linesOverlaySeries = linesOverlaySeries && line.overlaySeries;
-      }
+    for (final line in lines) {
+      linesOverlaySeries = linesOverlaySeries && line.overlaySeries;
     }
-
+  
     var boundsOverlaySeries = true;
     if (bounds != null) {
       for (final bound in bounds!) {
