@@ -70,10 +70,8 @@ class PieChart<D> extends BaseChart<D> {
 
       final details = renderer.getExpandedDatumDetails(seriesDatum);
 
-      if (details != null) {
-        entries.add(details);
-      }
-    }
+      entries.add(details);
+        }
 
     return entries;
   }
