@@ -59,9 +59,7 @@ class CircleSectorPainter {
 
     final centerOffset = new Offset(center.x.toDouble(), center.y.toDouble());
 
-    final isFullCircle = startAngle != null &&
-        endAngle != null &&
-        endAngle - startAngle == 2 * pi;
+    final isFullCircle = endAngle - startAngle == 2 * pi;
 
     final midpointAngle = (endAngle + startAngle) / 2;
 
